Move all of the Proof stuff into its own module.
